refactoring: split code to files
