+++ packages/telemetry/index.ts
@@ -1,1 +1,3 @@
 export * from './telemetry'
+export * from './functions'
+export * from './types'
